+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Implement map in transaction kernel library (#1396).
 - Added shutdown configuration options to the `miden-proving-service` proxy (#1405).
 - [BREAKING] Implement transaction script arguments for the `TransactionScript` (#1406).
+- Implement Display for `NoteType` (#1420).
 
 ## 0.9.1 (2025-05-30)